requirements.txt
schedule
pytz
discord.py